Refactor metadata into it's own component; move summaries to sidebar
This commit refactors the metadata sidebar that exists in both the
catalog and item views into a single component. The summary tab was
removed and the colleciton summary information is instead placed
inside the metadata sidebar.
