+++ src/config.js
@@ -1,0 +1,7 @@
+export const CATALOG_URL =
+  process.env.CATALOG_URL ||
+      "https://raw.githubusercontent.com/cholmes/sample-stac/master/stac/catalog.json";
+
+export const STAC_VERSION =
+  process.env.STAC_VERSION ||
+      "0.9.0";
